Check method exists first before instantiating
A ReflectionException is thrown in the constructor for ReflectionMethod as detailed at https://www.php.net/manual/en/reflectionmethod.construct.php. The assumption is that null is returned  if ReflectionMethod cannot be instantiated. This amend checks whether the method exists, creates a ReflectionMethod instance if it does or null otherwise. Fixes #29465